SSSE3 decoder: don't use 16-bit operations: slow
Benchmarking shows that using a uint16_t here is slower than using a
uint32_t, because accessing a 16-bit register incurs a small penalty.

+++ lib/dec/ssse3.c
@@ -40,7 +40,7 @@ while (srclen >= 24)
 	ormask = _mm_or_si128(ormask, s4mask);
 	ormask = _mm_or_si128(ormask, s5mask);
 
-	if ((uint16_t)_mm_movemask_epi8(ormask) != (uint16_t)0xFFFF) {
+	if ((uint32_t)_mm_movemask_epi8(ormask) != (uint32_t)0xFFFF) {
 		break;
 	}
